Update Stream model and add AppStore screenshot generator
Refactored the Stream model to replace `sources` with a single `source` and updated related logic in `CreatorView`. Added a new `AppStoreScreenshotGenerator` class to automate screenshot capture for App Store assets. Removed deprecated ChatView model code for cleanup.
