+++ Arkavo/Arkavo/Arkavo.entitlements
@@ -9,23 +9,7 @@
 		<string>applinks:arkavo.net</string>
 		<string>webcredentials:arkavo.net</string>
 	</array>
-	<key>com.apple.developer.authentication-services.autofill-credential-provider</key>
-	<true/>
-	<key>com.apple.security.app-sandbox</key>
-	<true/>
-	<key>com.apple.security.assets.movies.read-only</key>
-	<true/>
-	<key>com.apple.security.assets.music.read-only</key>
-	<true/>
-	<key>com.apple.security.assets.pictures.read-only</key>
-	<true/>
 	<key>com.apple.security.device.camera</key>
-	<true/>
-	<key>com.apple.security.files.downloads.read-only</key>
-	<true/>
-	<key>com.apple.security.files.user-selected.read-only</key>
-	<true/>
-	<key>com.apple.security.network.client</key>
 	<true/>
 	<key>keychain-access-groups</key>
 	<array>
