+++ Arkavo/Arkavo/ChatView.swift
@@ -4,578 +4,6 @@ import FlatBuffers
 import OpenTDFKit
 import SwiftUI
 
-// MARK: - View Model
-
-@MainActor
-class ChatViewModel: ObservableObject, ArkavoClientDelegate {
-    let client: ArkavoClient
-    let account: Account
-    let profile: Profile
-    let stream: Stream
-    @Published var messages: [ChatMessage] = []
-    @Published var connectionState: ArkavoClientState = .disconnected
-    // Track pending thoughts by their ephemeral public key
-    private var pendingThoughts: [Data: (header: Header, payload: Payload, nano: NanoTDF)] = [:]
-    // Add set to track processed message IDs
-    private var processedMessageIds = Set<Data>()
-    private var notificationObservers: [NSObjectProtocol] = []
-
-    init(client: ArkavoClient, account: Account, profile: Profile, stream: Stream) {
-        self.client = client
-        self.account = account
-        self.profile = profile
-        self.stream = stream
-        connectionState = client.currentState
-        setupNotifications()
-    }
-
-    private func setupNotifications() {
-        // Clean up any existing observers
-        notificationObservers.forEach { NotificationCenter.default.removeObserver($0) }
-        notificationObservers.removeAll()
-
-        // Add observers
-        let stateObserver = NotificationCenter.default.addObserver(
-            forName: .arkavoClientStateChanged,
-            object: nil,
-            queue: .main
-        ) { [weak self] notification in
-            guard let state = notification.userInfo?["state"] as? ArkavoClientState else { return }
-            Task { @MainActor in
-                self?.connectionState = state
-            }
-        }
-        notificationObservers.append(stateObserver)
-
-        let natsObserver = NotificationCenter.default.addObserver(
-            forName: .natsMessageReceived,
-            object: nil,
-            queue: .main
-        ) { [weak self] notification in
-            guard let data = notification.userInfo?["data"] as? Data else { return }
-            Task {
-                await self?.handleNATSMessage(data)
-            }
-        }
-        notificationObservers.append(natsObserver)
-
-        let keyObserver = NotificationCenter.default.addObserver(
-            forName: .rewrappedKeyReceived,
-            object: nil,
-            queue: .main
-        ) { [weak self] notification in
-            guard let data = notification.userInfo?["data"] as? Data else { return }
-            Task {
-                await self?.handleRewrappedKeyMessage(data)
-            }
-        }
-        notificationObservers.append(keyObserver)
-    }
-
-    deinit {
-        notificationObservers.forEach { NotificationCenter.default.removeObserver($0) }
-    }
-
-    func sendMessage(content: String) async throws {
-        let messageData = content.data(using: .utf8) ?? Data()
-        let streamPublicID = stream.publicID
-
-        // Create thought service model
-        let thoughtModel = ThoughtServiceModel(
-            creatorPublicID: profile.publicID,
-            streamPublicID: streamPublicID,
-            mediaType: .text,
-            content: messageData
-        )
-
-        // Create FlatBuffers policy
-        var builder = FlatBufferBuilder()
-
-        // Create format info
-        let formatVersionString = builder.create(string: "1.0")
-        let formatProfileString = builder.create(string: "standard")
-        let formatInfo = Arkavo_FormatInfo.createFormatInfo(
-            &builder,
-            type: .plain,
-            versionOffset: formatVersionString,
-            profileOffset: formatProfileString
-        )
-
-        // Create content format
-        let contentFormat = Arkavo_ContentFormat.createContentFormat(
-            &builder,
-            mediaType: .text,
-            dataEncoding: .utf8,
-            formatOffset: formatInfo
-        )
-
-        // Create rating based on stream age policy
-        let rating: Offset = switch stream.policies.age {
-        case .onlyAdults:
-            Arkavo_Rating.createRating(
-                &builder,
-                violent: .severe,
-                sexual: .severe,
-                profane: .severe,
-                substance: .severe,
-                hate: .severe,
-                harm: .severe,
-                mature: .severe,
-                bully: .severe
-            )
-        case .onlyKids:
-            Arkavo_Rating.createRating(
-                &builder,
-                violent: .mild,
-                sexual: .none_,
-                profane: .none_,
-                substance: .none_,
-                hate: .none_,
-                harm: .none_,
-                mature: .none_,
-                bully: .none_
-            )
-        case .forAll:
-            Arkavo_Rating.createRating(
-                &builder,
-                violent: .mild,
-                sexual: .mild,
-                profane: .mild,
-                substance: .none_,
-                hate: .none_,
-                harm: .none_,
-                mature: .mild,
-                bully: .none_
-            )
-        case .onlyTeens:
-            Arkavo_Rating.createRating(
-                &builder,
-                violent: .mild,
-                sexual: .none_,
-                profane: .none_,
-                substance: .none_,
-                hate: .none_,
-                harm: .none_,
-                mature: .none_,
-                bully: .none_
-            )
-        }
-
-        // Create purpose
-        let purpose = Arkavo_Purpose.createPurpose(
-            &builder,
-            educational: 0.8,
-            entertainment: 0.2,
-            news: 0.0,
-            promotional: 0.0,
-            personal: 0.0,
-            opinion: 0.1,
-            transactional: 0.0,
-            harmful: 0.0,
-            confidence: 0.9
-        )
-
-        // Create ID and related vectors
-        let idVector = builder.createVector(bytes: thoughtModel.publicID)
-        let relatedVector = builder.createVector(bytes: streamPublicID)
-
-        // Create topics vector
-        let topics: [UInt32] = [1, 2, 3]
-        let topicsVector = builder.createVector(topics)
-
-        // Create metadata root
-        let metadata = Arkavo_Metadata.createMetadata(
-            &builder,
-            created: Int64(Date().timeIntervalSince1970),
-            idVectorOffset: idVector,
-            relatedVectorOffset: relatedVector,
-            ratingOffset: rating,
-            purposeOffset: purpose,
-            topicsVectorOffset: topicsVector,
-            contentOffset: contentFormat
-        )
-
-        builder.finish(offset: metadata)
-
-        // Verify FlatBuffer
-        var buffer = builder.sizedBuffer
-        let rootOffset = buffer.read(def: Int32.self, position: 0)
-        var verifier = try Verifier(buffer: &buffer)
-        try Arkavo_Metadata.verify(&verifier, at: Int(rootOffset), of: Arkavo_Metadata.self)
-
-        // Get policy data
-        let policyData = Data(
-            bytes: buffer.memory.advanced(by: buffer.reader),
-            count: Int(buffer.size)
-        )
-
-        // Serialize payload
-        let payload = try thoughtModel.serialize()
-
-        // Encrypt and send via client
-        let nanoData = try await client.encryptAndSendPayload(
-            payload: payload,
-            policyData: policyData
-        )
-
-        let thoughtMetadata = Thought.Metadata(
-            creator: profile.id,
-            streamPublicID: streamPublicID,
-            mediaType: .text,
-            createdAt: Date(),
-            summary: content,
-            contributors: []
-        )
-
-        let thought = Thought(
-            nano: nanoData,
-            metadata: thoughtMetadata
-        )
-        _ = try PersistenceController.shared.saveThought(thought)
-        // Save thought to stream
-        stream.thoughts.append(thought)
-        try await PersistenceController.shared.saveChanges()
-        // Create and add local message
-        let message = ChatMessage(
-            id: UUID().uuidString,
-            userId: profile.id.uuidString,
-            username: profile.name,
-            content: content,
-            timestamp: Date(),
-            attachments: [],
-            reactions: [],
-            isPinned: false,
-            publicID: thought.publicID,
-            creatorPublicID: profile.publicID,
-            mediaType: .text,
-            rawContent: messageData
-        )
-
-        await MainActor.run {
-            messages.append(message)
-        }
-    }
-
-    func handleIncomingMessage(_ data: Data) async {
-        print("\n=== handleIncomingMessage ===")
-        guard let messageType = data.first else {
-            print("No message type byte found")
-            return
-        }
-        print("Message type: 0x\(String(format: "%02X", messageType))")
-
-        let messageData = data.dropFirst()
-
-        switch messageType {
-        case 0x04: // Rewrapped key
-            print("Routing to handleRewrappedKeyMessage")
-            await handleRewrappedKeyMessage(messageData)
-
-        case 0x05, 0x06: // NATS message/event
-            print("Routing to NATS message handler")
-            Task {
-                await handleNATSMessage(messageData)
-            }
-
-        default:
-            print("Unknown message type: 0x\(String(format: "%02X", messageType))")
-        }
-    }
-
-    private func handleRewrappedKeyMessage(_ data: Data) async {
-        print("\n=== handleRewrappedKeyMessage ===")
-        print("Message length: \(data.count)")
-
-        guard data.count == 93 else {
-            if data.count == 33 {
-                let identifier = data
-                print("Received DENY for EPK: \(identifier.hexEncodedString())")
-//                if let removedThought = pendingThoughts.removeValue(forKey: identifier) {
-//                    print("Removed denied thought from pending queue")
-//                    print("Remaining pending thoughts: \(pendingThoughts.count)")
-//                } else {
-//                    print("No matching thought found for denied EPK")
-//                }
-                return
-            }
-            print("Invalid rewrapped key length: \(data.count)")
-            return
-        }
-
-        let identifier = data.prefix(33)
-        print("Looking for thought with EPK: \(identifier.hexEncodedString())")
-        print("Current pending thoughts: \(pendingThoughts.keys.map { $0.hexEncodedString() })")
-
-        // Find corresponding thought
-        guard let (header, _, nano) = pendingThoughts.removeValue(forKey: identifier) else {
-            print("❌ No pending thought found for EPK: \(identifier.hexEncodedString())")
-            return
-        }
-
-        print("✅ Found matching thought!")
-        let keyData = data.suffix(60)
-        let nonce = keyData.prefix(12)
-        let encryptedKeyLength = keyData.count - 12 - 16
-        let rewrappedKey = keyData.prefix(keyData.count - 16).suffix(encryptedKeyLength)
-        let authTag = keyData.suffix(16)
-
-        do {
-            print("Attempting to decrypt rewrapped key...")
-            let symmetricKey = try client.decryptRewrappedKey(
-                nonce: nonce,
-                rewrappedKey: rewrappedKey,
-                authTag: authTag
-            )
-            print("Successfully decrypted rewrapped key")
-
-            // Decrypt the thought payload
-            print("Attempting to decrypt thought payload...")
-            let decryptedData = try await nano.getPayloadPlaintext(symmetricKey: symmetricKey)
-            print("Successfully decrypted thought payload")
-
-            // Process decrypted thought
-            Task {
-                print("Processing decrypted thought...")
-                await handleDecryptedThought(
-                    payload: decryptedData,
-                    policy: ArkavoPolicy(header.policy),
-                    nano: nano
-                )
-            }
-        } catch {
-            print("❌ Error processing rewrapped key: \(error)")
-        }
-    }
-
-    private func handleNATSMessage(_ data: Data) async {
-        do {
-            // Create a deep copy of the data
-            let copiedData = Data(data)
-            let parser = BinaryParser(data: copiedData)
-            let header = try parser.parseHeader()
-            print("\n=== Processing NATS Message ===")
-            print("Checking content rating...")
-            print("Message content rating: \(header.policy)")
-            let payload = try parser.parsePayload(config: header.payloadSignatureConfig)
-            let nano = NanoTDF(header: header, payload: payload, signature: nil)
-
-            let epk = header.ephemeralPublicKey
-            print("Parsed NATS message - EPK: \(epk.hexEncodedString())")
-
-            // Store with correct types
-            pendingThoughts[epk] = (
-                header: header,
-                payload: payload,
-                nano: nano
-            )
-            print("Current pending thoughts count: \(pendingThoughts.count)")
-            print("Pending thoughts EPKs: \(pendingThoughts.keys.map { $0.hexEncodedString() })")
-            // Send rewrap message
-            print("Sending rewrap message:")
-            print("EPK: \(header.ephemeralPublicKey.hexEncodedString())")
-            let rewrapMessage = RewrapMessage(header: header)
-            let messageData = rewrapMessage.toData()
-            print("Rewrap message first byte: 0x\(String(format: "%02X", messageData.first ?? 0))")
-            print("Rewrap message length: \(messageData.count)")
-            print("Sent rewrap message for EPK: \(epk.hexEncodedString())")
-        } catch {
-            print("Error processing NATS message: \(error)")
-        }
-    }
-
-    func processStreamThoughts() async {
-        print("\nProcessing stream thoughts:")
-        for thought in stream.thoughts {
-            do {
-                print("\n=== Processing thought ===")
-                print("Public ID: \(thought.publicID.hexEncodedString())")
-                print("Nano data size: \(thought.nano.count) bytes")
-
-                // Debug the first few bytes
-                let previewSize = min(thought.nano.count, 8)
-                let previewBytes = thought.nano.prefix(previewSize)
-                print("First \(previewSize) bytes: \(previewBytes.map { String(format: "%02X", $0) }.joined(separator: " "))")
-
-                // Check expected magic number
-                print("Expected magic number: \(Header.magicNumber.map { String(format: "%02X", $0) }.joined(separator: " "))")
-
-                print("Creating binary parser...")
-                let parser = BinaryParser(data: thought.nano)
-
-                print("Attempting to parse header...")
-                let header = try parser.parseHeader()
-                print("✅ Successfully parsed header")
-                print(" - EPK length: \(header.ephemeralPublicKey.count)")
-                print(" - EPK: \(header.ephemeralPublicKey.hexEncodedString())")
-
-                print("Attempting to parse payload...")
-                let payload = try parser.parsePayload(config: header.payloadSignatureConfig)
-                print("✅ Successfully parsed payload:")
-                print(" - Payload length: \(payload.length)")
-                print(" - IV length: \(payload.iv.count)")
-                print(" - Ciphertext length: \(payload.ciphertext.count)")
-                print(" - MAC length: \(payload.mac.count)")
-
-                let nano = NanoTDF(header: header, payload: payload, signature: nil)
-
-                // Store in pending thoughts
-                pendingThoughts[header.ephemeralPublicKey] = (header, payload, nano)
-                print("Added to pending thoughts with EPK: \(header.ephemeralPublicKey.hexEncodedString())")
-
-                // Send rewrap message
-                let rewrapMessage = RewrapMessage(header: header)
-                try await client.sendMessage(rewrapMessage.toData())
-                print("✅ Sent rewrap message")
-
-            } catch {
-                print("❌ Error processing thought: \(error)")
-                if let parsingError = error as? ParsingError {
-                    switch parsingError {
-                    case .invalidMagicNumber:
-                        print("Magic number check failed - first bytes don't match 4C 31")
-                        print("Expected: 4C 31")
-                        if thought.nano.count >= 2 {
-                            print("Found: \(thought.nano.prefix(2).map { String(format: "%02X", $0) }.joined(separator: " "))")
-                        }
-                    case .invalidFormat:
-                        print("Invalid format - data structure doesn't match expected TDF format")
-                        print("This could be due to:")
-                        print("- Incorrect field lengths")
-                        print("- Missing required fields")
-                        print("- Malformed data")
-                    default:
-                        print("Other parsing error: \(parsingError)")
-                    }
-                }
-            }
-        }
-    }
-
-    func handleStreamData(_ data: Data) async throws {
-        print("Handling stream data of length: \(data.count)")
-
-        var buffer = ByteBuffer(data: data)
-        print("Created ByteBuffer")
-
-        // Add try-catch for better error handling
-        do {
-            let rootOffset = buffer.read(def: Int32.self, position: 0)
-            print("Root offset: \(rootOffset)")
-
-            var verifier = try Verifier(buffer: &buffer)
-            try Arkavo_EntityRoot.verify(&verifier, at: Int(rootOffset), of: Arkavo_EntityRoot.self)
-            print("Verification passed")
-
-            // ... rest of the processing ...
-
-            print("Stream processing complete")
-        } catch {
-            print("Error processing stream data: \(error)")
-            throw error
-        }
-    }
-
-    private func handleDecryptedThought(payload: Data, policy _: ArkavoPolicy, nano _: NanoTDF) async {
-        print("\nHandling decrypted thought:")
-        do {
-            let thoughtModel = try ThoughtServiceModel.deserialize(from: payload)
-            print("Deserialized thought model")
-            print("Media type: \(thoughtModel.mediaType)")
-            // Check if we've already processed this message
-            if processedMessageIds.contains(thoughtModel.publicID) {
-                print("⚠️ Skipping duplicate message with ID: \(thoughtModel.publicID.hexEncodedString())")
-                return
-            }
-            // Process content based on media type
-            let displayContent = processContent(thoughtModel.content, mediaType: thoughtModel.mediaType)
-            print("Processed content: \(displayContent)")
-
-            let message = ChatMessage(
-                id: UUID().uuidString,
-                userId: thoughtModel.creatorPublicID.base58EncodedString,
-                username: formatUsername(publicID: thoughtModel.creatorPublicID),
-                content: displayContent,
-                timestamp: Date(),
-                attachments: [],
-                reactions: [],
-                isPinned: false,
-                publicID: thoughtModel.publicID,
-                creatorPublicID: thoughtModel.creatorPublicID,
-                mediaType: thoughtModel.mediaType,
-                rawContent: thoughtModel.content // Store original content
-            )
-
-            await MainActor.run {
-                messages.append(message)
-                // Add this line to prevent reprocessing:
-                processedMessageIds.insert(thoughtModel.publicID)
-                print("✅ Added message to chat view - Type: \(thoughtModel.mediaType)")
-
-                // Debug current messages
-                print("Current messages:")
-                for (index, msg) in messages.enumerated() {
-                    print("[\(index)] \(msg.username) [\(msg.mediaType)]: \(msg.content)")
-                }
-            }
-        } catch {
-            print("❌ Error handling decrypted thought: \(error)")
-        }
-    }
-
-    private func processContent(_ content: Data, mediaType: MediaType) -> String {
-        switch mediaType {
-        case .text:
-            return String(data: content, encoding: .utf8) ?? "[Invalid text content]"
-
-        case .image:
-            if UIImage(data: content) != nil {
-                return "[Image]"
-            }
-            return "[Invalid image data]"
-
-        case .video:
-            return "[Video content]"
-
-        case .audio:
-            return "[Audio content]"
-        }
-    }
-
-    private func formatUsername(publicID: Data) -> String {
-        let shortID = publicID.prefix(6).base58EncodedString
-        return "User-\(shortID)"
-    }
-
-    enum ChatError: Error {
-        case noProfile
-        case serializationError
-    }
-
-    nonisolated func clientDidChangeState(_: ArkavoClient, state: ArkavoClientState) {
-        Task { @MainActor in
-            self.connectionState = state
-        }
-    }
-
-    nonisolated func clientDidReceiveMessage(_: ArkavoClient, message: Data) {
-        Task { @MainActor in
-            print("\n=== clientDidReceiveMessage ===")
-            await self.handleIncomingMessage(message)
-        }
-    }
-
-    nonisolated func clientDidReceiveError(_: ArkavoClient, error: Error) {
-        Task { @MainActor in
-            print("Arkavo client error: \(error)")
-            // You might want to update UI state here
-            // self.errorMessage = error.localizedDescription
-            // self.showError = true
-        }
-    }
-}
-
-// MARK: - Enhanced ChatView
-
 struct ChatView: View {
     @StateObject var viewModel: ChatViewModel
     @State private var messageText = ""
